Rename files to reflect get_startgg command.
